+++ src/functionalTest/groovy/com/autonomousapps/android/AbstractAndroidSpec.groovy
@@ -25,7 +25,7 @@ abstract class AbstractAndroidSpec extends AbstractFunctionalSpec {
   protected static final AGP_7_2 = AgpVersion.version('7.2.0-beta04')
   protected static final AGP_7_3 = AgpVersion.version('7.3.0-alpha08')
 
-  private static final SUPPORTED_AGP_VERSIONS = [
+  protected static final SUPPORTED_AGP_VERSIONS = [
     AGP_4_2,
 //    AGP_7_0,
     AGP_7_1,