+++ src/functionalTest/groovy/com/autonomousapps/android/ExternalApplicationSpec.groovy
@@ -3,7 +3,9 @@ package com.autonomousapps.android
 import com.autonomousapps.android.projects.ExternalApplicationProject
 import org.gradle.util.GradleVersion
 
+import static com.autonomousapps.advice.truth.BuildHealthSubject.buildHealth
 import static com.autonomousapps.utils.Runner.build
+import static com.google.common.truth.Truth.assertAbout
 import static com.google.common.truth.Truth.assertThat
 
 final class ExternalApplicationSpec extends AbstractAndroidSpec {
@@ -17,7 +19,9 @@ final class ExternalApplicationSpec extends AbstractAndroidSpec {
     build(gradleVersion as GradleVersion, gradleProject.rootDir, 'buildHealth')
 
     then:
-    assertThat(project.actualBuildHealth()).containsExactlyElementsIn(project.expectedBuildHealth)
+    assertAbout(buildHealth())
+      .that(project.actualBuildHealth())
+      .isEquivalentIgnoringModuleAdvice(project.expectedBuildHealth)
 
     where:
     [gradleVersion, agpVersion] << gradleAgpMatrix()