+++ src/main/kotlin/com/autonomousapps/model/ProjectVariant.kt
@@ -3,9 +3,11 @@ package com.autonomousapps.model
 import com.autonomousapps.internal.unsafeLazy
 import com.autonomousapps.internal.utils.flatMapToOrderedSet
 import com.autonomousapps.internal.utils.flatMapToSet
+import com.autonomousapps.internal.utils.fromJson
 import com.autonomousapps.model.CodeSource.Kind
 import com.autonomousapps.model.declaration.Variant
 import com.squareup.moshi.JsonClass
+import org.gradle.api.file.Directory
 
 /** Represents a variant-specific view of the project under analysis. */
 @Suppress("MemberVisibilityCanBePrivate") // deliberate API
@@ -54,6 +56,14 @@ data class ProjectVariant(
     sources.filterIsInstance<AndroidResSource>()
   }
 
+  val androidAssetsSource: List<AndroidAssetSource> by unsafeLazy {
+    sources.filterIsInstance<AndroidAssetSource>()
+  }
+
+  val imports: Set<String> by unsafeLazy {
+    codeSource.flatMapToOrderedSet { it.imports }
+  }
+
   val javaImports: Set<String> by unsafeLazy {
     codeSource.filter { it.kind == Kind.JAVA }
       .flatMapToOrderedSet { it.imports }
@@ -74,12 +84,17 @@ data class ProjectVariant(
       .flatMapToOrderedSet { it.imports }
   }
 
-  val imports: Set<String> by unsafeLazy {
-    sortedSetOf<String>().apply {
-      addAll(javaImports)
-      addAll(kotlinImports)
-      addAll(groovyImports)
-      addAll(scalaImports)
-    }
+  internal fun dependencies(dependenciesDir: Directory): Set<Dependency> {
+    return classpath.asSequence()
+      .plus(annotationProcessors)
+      .map {
+        val file = dependenciesDir.file(it.toFileName())
+        if (file.asFile.exists()) {
+          file.fromJson<Dependency>()
+        } else {
+          error("No file for ${coordinates.gav()}")
+        }
+      }
+      .toSet()
   }
 }